fix(sync): widen tree op position to long to stop append-sentinel parse failure
TreeAdd.position / TreeMove.newPosition were int, but the frontend
appends with a Number.MAX_SAFE_INTEGER (9_007_199_254_740_991) sentinel
meaning "end of the sibling list". That value overflows a Java int, so
Jackson rejected the whole frame at readValue() and the handler's
parse-catch swallowed it — no ack, no conflict, the tree mutation never
applied. Sheet/folder seed data hid this because it is written to
sheet_tree directly; the first live tree.add (a new document) was the
first frame to hit the op path, so doc creation always 404'd on the
collab token (the paired documents row was never inserted).

Widen both fields to long and narrow the clamped index back to int at
the two insert() sites (the clamp guarantees it fits).

+++ packages/backend/src/main/java/com/balruno/sync/SyncMessage.java
@@ -187,23 +187,31 @@ public sealed interface SyncMessage {
         }
     }
 
+    // position / newPosition are {@code long}, not {@code int}: the
+    // frontend appends with a {@code Number.MAX_SAFE_INTEGER}
+    // (9_007_199_254_740_991) sentinel meaning "end of the sibling
+    // list", which the server clamps to {@code [0, size]}. That value
+    // overflows a Java int, so an int field makes Jackson reject the
+    // whole frame at readValue() time — the op is then swallowed by the
+    // handler's parse-catch with no ack and no conflict, and the tree
+    // mutation silently never happens (the doc-never-persists bug).
     record TreeAdd(
-            TreeKind treeKind, UUID parentId, int position, Object node,
+            TreeKind treeKind, UUID parentId, long position, Object node,
             long baseVersion, UUID clientMsgId,
             UndoMeta undo
     ) implements SyncMessage {
-        public TreeAdd(TreeKind treeKind, UUID parentId, int position, Object node,
+        public TreeAdd(TreeKind treeKind, UUID parentId, long position, Object node,
                        long baseVersion, UUID clientMsgId) {
             this(treeKind, parentId, position, node, baseVersion, clientMsgId, null);
         }
     }
 
     record TreeMove(
-            TreeKind treeKind, UUID nodeId, UUID newParentId, int newPosition,
+            TreeKind treeKind, UUID nodeId, UUID newParentId, long newPosition,
             long baseVersion, UUID clientMsgId,
             UndoMeta undo
     ) implements SyncMessage {
-        public TreeMove(TreeKind treeKind, UUID nodeId, UUID newParentId, int newPosition,
+        public TreeMove(TreeKind treeKind, UUID nodeId, UUID newParentId, long newPosition,
                         long baseVersion, UUID clientMsgId) {
             this(treeKind, nodeId, newParentId, newPosition, baseVersion, clientMsgId, null);
         }
